Refactor code and delete useless files
