+++ conf/train/default.yaml
@@ -1,17 +1,16 @@
 # reproducibility
 seed_index: 0
-deterministic: False
+deterministic: True
 
 # PyTorch Lightning Trainer https://pytorch-lightning.readthedocs.io/en/stable/common/trainer.html
 trainer:
   fast_dev_run: False # Enable this for debug purposes
   gpus: 1
   precision: 32
-  max_epochs: 3
-  max_steps: 10000
+  max_epochs: 1000
   accumulate_grad_batches: 1
   num_sanity_val_steps: 2
-  gradient_clip_val: 10.0
+  gradient_clip_val: 0
   val_check_interval: 1.0
   deterministic: ${train.deterministic}
