+++ src/vision_transformer/data/datamodule.py
@@ -1,76 +1,35 @@
 import logging
 from functools import cached_property, partial
-from pathlib import Path
-from typing import List, Mapping, Optional, Sequence, Union
-
+from collections import Counter
 import hydra
 import omegaconf
 import pytorch_lightning as pl
 from omegaconf import DictConfig
-from torch.utils.data import DataLoader, Dataset, random_split
-from torch.utils.data.dataloader import default_collate
-from torchvision import transforms
-
 from nn_core.common import PROJECT_ROOT
-from nn_core.nn_types import Split
+from nn_core.model_logging import NNLogger
+from torchvision import datasets, transforms
+from torch.utils.data import DataLoader, random_split
+from torch.utils.data.dataloader import SequentialSampler, default_collate
+
+from .sampler import RASampler
+from .augmentation import GaussianBlur, Solarization
 
 pylogger = logging.getLogger(__name__)
 
 
 class MetaData:
-    def __init__(self, class_vocab: Mapping[str, int]):
-        """The data information the Lightning Module will be provided with.
+    def __init__(self, class_vocab):
+        self.class_vocab = class_vocab
 
-        This is a "bridge" between the Lightning DataModule and the Lightning Module.
-        There is no constraint on the class name nor in the stored information, as long as it exposes the
-        `save` and `load` methods.
-
-        The Lightning Module will receive an instance of MetaData when instantiated,
-        both in the train loop or when restored from a checkpoint.
-
-        This decoupling allows the architecture to be parametric (e.g. in the number of classes) and
-        DataModule/Trainer independent (useful in prediction scenarios).
-        MetaData should contain all the information needed at test time, derived from its train dataset.
-
-        Examples are the class names in a classification task or the vocabulary in NLP tasks.
-        MetaData exposes `save` and `load`. Those are two user-defined methods that specify
-        how to serialize and de-serialize the information contained in its attributes.
-        This is needed for the checkpointing restore to work properly.
-
-        Args:
-            class_vocab: association between class names and their indices
-        """
-        # example
-        self.class_vocab: Mapping[str, int] = class_vocab
-
-    def save(self, dst_path: Path) -> None:
-        """Serialize the MetaData attributes into the zipped checkpoint in dst_path.
-
-        Args:
-            dst_path: the root folder of the metadata inside the zipped checkpoint
-        """
-        pylogger.debug(f"Saving MetaData to '{dst_path}'")
-
-        # example
+    def save(self, dst_path):
+        pylogger.debug(f"Saving Metadata to {dst_path}")
         (dst_path / "class_vocab.tsv").write_text(
             "\n".join(f"{key}\t{value}" for key, value in self.class_vocab.items())
         )
 
-    @staticmethod
-    def load(src_path: Path) -> "MetaData":
-        """Deserialize the MetaData from the information contained inside the zipped checkpoint in src_path.
-
-        Args:
-            src_path: the root folder of the metadata inside the zipped checkpoint
-
-        Returns:
-            an instance of MetaData containing the information in the checkpoint
-        """
-        pylogger.debug(f"Loading MetaData from '{src_path}'")
-
-        # example
+    def load(self, src_path):
+        pylogger.debug(f"Loading MetaData from {src_path}")
         lines = (src_path / "class_vocab.tsv").read_text(encoding="utf-8").splitlines()
-
         class_vocab = {}
         for line in lines:
             key, value = line.strip().split("\t")
@@ -81,140 +40,128 @@ class MetaData:
         )
 
 
-def collate_fn(samples: List, split: Split, metadata: MetaData):
-    """Custom collate function for dataloaders with access to split and metadata.
-
-    Args:
-        samples: A list of samples coming from the Dataset to be merged into a batch
-        split: The data split (e.g. train/val/test)
-        metadata: The MetaData instance coming from the DataModule or the restored checkpoint
-
-    Returns:
-        A batch generated from the given samples
-    """
+def collate_fn(samples, split, metadata):
     return default_collate(samples)
 
 
 class MyDataModule(pl.LightningDataModule):
     def __init__(
         self,
-        datasets: DictConfig,
-        num_workers: DictConfig,
-        batch_size: DictConfig,
-        gpus: Optional[Union[List[int], str, int]],
-        # example
-        val_percentage: float,
+        datasets,
+        num_workers,
+        batch_size,
+        gpus,
+        val_percentage,
     ):
         super().__init__()
         self.datasets = datasets
         self.num_workers = num_workers
         self.batch_size = batch_size
-        # https://pytorch-lightning.readthedocs.io/en/stable/common/trainer.html#gpus
-        self.pin_memory: bool = gpus is not None and str(gpus) != "0"
-
-        self.train_dataset: Optional[Dataset] = None
-        self.val_datasets: Optional[Sequence[Dataset]] = None
-        self.test_datasets: Optional[Sequence[Dataset]] = None
-
-        # example
-        self.val_percentage: float = val_percentage
+        self.gpus = gpus
+        self.pin_memory = gpus is not None and str(gpus) != "0"
+        self.train_dataset = None
+        self.val_dataset = None
+        self.test_dataset = None
+        self.val_percentage = val_percentage
 
     @cached_property
-    def metadata(self) -> MetaData:
-        """Data information to be fed to the Lightning Module as parameter.
-
-        Examples are vocabularies, number of classes...
-
-        Returns:
-            metadata: everything the model should know about the data, wrapped in a MetaData object.
-        """
-        # Since MetaData depends on the training data, we need to ensure the setup method has been called.
+    def metadata(self):
         if self.train_dataset is None:
             self.setup(stage="fit")
-
         return MetaData(class_vocab=self.train_dataset.dataset.class_vocab)
 
-    def prepare_data(self) -> None:
-        # download only
-        pass
+    def setup(self, stage):
+        train_transform = transforms.Compose(
+            [
+                transforms.Resize(224, interpolation=3),
+                transforms.RandomCrop(224, padding=4,padding_mode='reflect'),
+                transforms.RandomHorizontalFlip(),
+                transforms.RandomChoice([Solarization(),GaussianBlur()]),
+                transforms.ColorJitter(0.3,0.3,0.3),
+                transforms.ToTensor(),
+                transforms.Normalize((0.4914, 0.4822, 0.4465), (0.247, 0.243, 0.261)),
+            ]
+        )
+        test_transform = transforms.Compose(
+            [
+                transforms.Resize(224, interpolation=3),
+                transforms.ToTensor(),
+                transforms.Normalize((0.4914, 0.4822, 0.4465), (0.247, 0.243, 0.261)),
+            ]
+        )
 
-    def setup(self, stage: Optional[str] = None):
-        transform = transforms.Compose([transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))])
-
-        # Here you should instantiate your datasets, you may also split the train into train and validation if needed.
-        if (stage is None or stage == "fit") and (self.train_dataset is None and self.val_datasets is None):
-            # example
-            mnist_train = hydra.utils.instantiate(
+        if (stage is None or stage == "fit") and (self.train_dataset is None and self.val_dataset is None):
+            cifar10_train = hydra.utils.instantiate(
                 self.datasets.train,
                 split="train",
-                transform=transform,
+                transform=train_transform,
                 path=PROJECT_ROOT / "data",
             )
-            train_length = int(len(mnist_train) * (1 - self.val_percentage))
-            val_length = len(mnist_train) - train_length
-            self.train_dataset, val_dataset = random_split(mnist_train, [train_length, val_length])
+            train_length = int(len(cifar10_train) * (1 - self.val_percentage))
+            val_length = len(cifar10_train) - train_length
+            self.train_dataset, self.val_dataset = random_split(cifar10_train, [train_length, val_length])
+            self.val_dataset.transforms = test_transform
 
-            self.val_datasets = [val_dataset]
+            print("ECCCOMI QUA MANNAGGIA LA PUTTANA")
 
         if stage is None or stage == "test":
-            self.test_datasets = [
-                hydra.utils.instantiate(
-                    dataset_cfg,
-                    split="test",
-                    path=PROJECT_ROOT / "data",
-                    transform=transform,
-                )
-                for dataset_cfg in self.datasets.test
-            ]
+            self.test_dataset = hydra.utils.instantiate(
+                self.datasets.test,
+                split="test",
+                path=PROJECT_ROOT / "data",
+                transform=test_transform,
+            )
 
-    def train_dataloader(self) -> DataLoader:
+    def train_dataloader(self):
         return DataLoader(
             self.train_dataset,
-            shuffle=True,
             batch_size=self.batch_size.train,
             num_workers=self.num_workers.train,
             pin_memory=self.pin_memory,
-            collate_fn=partial(collate_fn, split="train", metadata=self.metadata),
+            collate_fn=partial(collate_fn, split="train", metadata=self.metadata)
         )
 
-    def val_dataloader(self) -> Sequence[DataLoader]:
-        return [
-            DataLoader(
-                dataset,
-                shuffle=False,
-                batch_size=self.batch_size.val,
-                num_workers=self.num_workers.val,
-                pin_memory=self.pin_memory,
-                collate_fn=partial(collate_fn, split="val", metadata=self.metadata),
-            )
-            for dataset in self.val_datasets
-        ]
+    def val_dataloader(self):
+        return DataLoader(
+            self.val_dataset,
+            shuffle=False,
+            batch_size=self.batch_size.val,
+            num_workers=self.num_workers.val,
+            pin_memory=self.pin_memory,
+            collate_fn=partial(collate_fn, split="val", metadata=self.metadata)
+        )
 
-    def test_dataloader(self) -> Sequence[DataLoader]:
-        return [
-            DataLoader(
-                dataset,
-                shuffle=False,
-                batch_size=self.batch_size.test,
-                num_workers=self.num_workers.test,
-                pin_memory=self.pin_memory,
-                collate_fn=partial(collate_fn, split="test", metadata=self.metadata),
-            )
-            for dataset in self.test_datasets
-        ]
+    def test_dataloader(self):
+        return DataLoader(
+            self.test_dataset,
+            shuffle=False,
+            batch_size=self.batch_size.test,
+            num_workers=self.num_workers.test,
+            pin_memory=self.pin_memory,
+            collate_fn=partial(collate_fn, split="test", metadata=self.metadata)
+        )
 
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}(" f"{self.datasets=}, " f"{self.num_workers=}, " f"{self.batch_size=})"
 
 
-@hydra.main(config_path=str(PROJECT_ROOT / "conf"), config_name="default")
-def main(cfg: omegaconf.DictConfig) -> None:
-    """Debug main to quickly develop the DataModule.
-
-    Args:
-        cfg: the hydra configuration
-    """
+@hydra.main(config_path=str(PROJECT_ROOT / "conf"), config_name="default", version_base=None)
+def main(cfg):
     _: pl.LightningDataModule = hydra.utils.instantiate(cfg.data.datamodule, _recursive_=False)
+    # Cound label distributions
+    _.setup("")
+    train_counter = Counter()
+    for images, labels in _.train_dataloader():
+        train_counter.update(labels.tolist())
+    print(f"Training label distribution\n{sorted(train_counter.items())}")
+    val_counter = Counter()
+    for images, labels in _.val_dataloader():
+        val_counter.update(labels.tolist())
+    print(f"Validation label distribution\n{sorted(val_counter.items())}")
+    test_counter = Counter()
+    for images, labels in _.test_dataloader():
+        test_counter.update(labels.tolist())
+    print(f"Testing label distribution\n{sorted(test_counter.items())}")
 
 
 if __name__ == "__main__":
