+++ src/vision_transformer/data/dataset.py
@@ -1,7 +1,7 @@
 import hydra
 import omegaconf
 from torch.utils.data import Dataset
-from torchvision.datasets import FashionMNIST
+from torchvision.datasets import CIFAR10
 
 from nn_core.common import PROJECT_ROOT
 from nn_core.nn_types import Split
@@ -13,7 +13,7 @@ class MyDataset(Dataset):
         self.split: Split = split
 
         # example
-        self.mnist = FashionMNIST(
+        self.cifar10 = CIFAR10(
             kwargs["path"],
             train=split == "train",
             download=True,
@@ -22,21 +22,21 @@ class MyDataset(Dataset):
 
     @property
     def class_vocab(self):
-        return self.mnist.class_to_idx
+        return self.cifar10.class_to_idx
 
     def __len__(self) -> int:
         # example
-        return len(self.mnist)
+        return len(self.cifar10)
 
     def __getitem__(self, index: int):
         # example
-        return self.mnist[index]
+        return self.cifar10[index]
 
     def __repr__(self) -> str:
         return f"MyDataset({self.split=}, n_instances={len(self)})"
 
 
-@hydra.main(config_path=str(PROJECT_ROOT / "conf"), config_name="default")
+@hydra.main(config_path=str(PROJECT_ROOT / "conf"), config_name="default", version_base=None)
 def main(cfg: omegaconf.DictConfig) -> None:
     """Debug main to quickly develop the Dataset.
