+++ src/vision_transformer/pl_modules/pl_module.py
@@ -1,66 +1,62 @@
 import logging
-from typing import Any, Mapping, Optional, Sequence, Tuple, Union
-
 import hydra
 import omegaconf
 import pytorch_lightning as pl
 import torch
 import torch.nn.functional as F
 import torchmetrics
-from torch.optim import Optimizer
-
 from nn_core.common import PROJECT_ROOT
 from nn_core.model_logging import NNLogger
+from torch.optim import Optimizer
+from timm.scheduler import CosineLRScheduler
+from timm.data import Mixup
+from timm.loss import SoftTargetCrossEntropy
 
 from vision_transformer.data.datamodule import MetaData
-from vision_transformer.modules.module import CNN
+from vision_transformer.modules.module import VisionTransformer
 
 pylogger = logging.getLogger(__name__)
 
 
-class MyLightningModule(pl.LightningModule):
+class LightningModel(pl.LightningModule):
     logger: NNLogger
 
-    def __init__(self, metadata: Optional[MetaData] = None, *args, **kwargs) -> None:
+    def __init__(self, metadata=None, *args, **kwargs):
         super().__init__()
-
-        # Populate self.hparams with args and kwargs automagically!
-        # We want to skip metadata since it is saved separately by the NNCheckpointIO object.
-        # Be careful when modifying this instruction. If in doubt, don't do it :]
         self.save_hyperparameters(logger=False, ignore=("metadata",))
 
         self.metadata = metadata
-
-        # example
         metric = torchmetrics.Accuracy()
         self.train_accuracy = metric.clone()
         self.val_accuracy = metric.clone()
         self.test_accuracy = metric.clone()
+        self.model = VisionTransformer(img_size=224,n_classes=10)
 
-        self.model = CNN(num_classes=len(metadata.class_vocab))
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """Method for the forward pass.
-
-        'training_step', 'validation_step' and 'test_step' should call
-        this method in order to compute the output predictions and the loss.
-
-        Returns:
-            output_dict: forward output containing the predictions (output logits ecc...) and the loss if any.
-        """
-        # example
+    def forward(self, x):
         return self.model(x)
 
-    def step(self, x, y) -> Mapping[str, Any]:
-        # example
-        logits = self(x)
-        loss = F.cross_entropy(logits, y)
+    def step(self, features, labels, train=None):
+        logits = self(features)
+        if train:
+            loss = SoftTargetCrossEntropy()(logits, labels)
+        else:
+            loss = F.cross_entropy(logits, labels)
         return {"logits": logits.detach(), "loss": loss}
 
-    def training_step(self, batch: Any, batch_idx: int) -> Mapping[str, Any]:
-        # example
-        x, y = batch
-        step_out = self.step(x, y)
+    def training_step(self, batch, batch_idx):
+        features, labels = batch
+        # mixup_fn = Mixup(
+        #     mixup_alpha=0.8,
+        #     cutmix_alpha=1.0,
+        #     cutmix_minmax=None,
+        #     prob=1,
+        #     switch_prob=0.5,
+        #     mode="batch",
+        #     num_classes=10,
+        #     label_smoothing=0.1,
+        # )
+        # features, smoothing_labels = mixup_fn(features, labels)
+        step_out = self.step(features, labels, train=False)
 
         self.log_dict(
             {"loss/train": step_out["loss"].cpu().detach()},
@@ -69,21 +65,19 @@ class MyLightningModule(pl.LightningModule):
             prog_bar=True,
         )
 
-        self.train_accuracy(torch.softmax(step_out["logits"], dim=-1), y)
+
+        self.train_accuracy(torch.softmax(step_out["logits"], dim=-1), labels)
         self.log_dict(
             {
                 "acc/train": self.train_accuracy,
             },
             on_epoch=True,
         )
-
         return step_out
 
-    def validation_step(self, batch: Any, batch_idx: int) -> Mapping[str, Any]:
-        # example
-        x, y = batch
-        step_out = self.step(x, y)
-
+    def validation_step(self, batch, batch_idx):
+        features, labels = batch
+        step_out = self.step(features, labels)
         self.log_dict(
             {"loss/val": step_out["loss"].cpu().detach()},
             on_step=False,
@@ -91,60 +85,62 @@ class MyLightningModule(pl.LightningModule):
             prog_bar=True,
         )
 
-        self.val_accuracy(torch.softmax(step_out["logits"], dim=-1), y)
+        self.val_accuracy(torch.softmax(step_out["logits"], dim=-1), labels)
         self.log_dict(
             {
                 "acc/val": self.val_accuracy,
             },
             on_epoch=True,
         )
-
         return step_out
 
-    def test_step(self, batch: Any, batch_idx: int) -> Mapping[str, Any]:
-        # example
-        x, y = batch
-        step_out = self.step(x, y)
+    def test_step(self, batch, batch_idx):
+        features, labels = batch
+        step_out = self.step(features, labels)
 
         self.log_dict(
             {"loss/test": step_out["loss"].cpu().detach()},
         )
 
-        self.test_accuracy(torch.softmax(step_out["logits"], dim=-1), y)
+        self.test_accuracy(torch.softmax(step_out["logits"], dim=-1), labels)
         self.log_dict(
             {
                 "acc/test": self.test_accuracy,
             },
             on_epoch=True,
         )
-
         return step_out
 
-    def configure_optimizers(
-        self,
-    ) -> Union[Optimizer, Tuple[Sequence[Optimizer], Sequence[Any]]]:
-        """Choose what optimizers and learning-rate schedulers to use in your optimization.
-
-        Normally you'd need one. But in the case of GANs or similar you might have multiple.
-
-        Return:
-            Any of these 6 options.
-            - Single optimizer.
-            - List or Tuple - List of optimizers.
-            - Two lists - The first list has multiple optimizers, the second a list of LR schedulers (or lr_dict).
-            - Dictionary, with an 'optimizer' key, and (optionally) a 'lr_scheduler'
-              key whose value is a single LR scheduler or lr_dict.
-            - Tuple of dictionaries as described, with an optional 'frequency' key.
-            - None - Fit will run without any optimizer.
-        """
+    def configure_optimizers_(self):
+        optimizer = torch.optim.AdamW(
+            self.parameters(),
+            lr=self.learning_rate,
+            weight_decay=0.05,
+        )
+        scheduler = CosineLRScheduler(
+            optimizer,
+            warmup_t=self.warmup_epochs,
+            t_initial=self.max_epochs,
+            lr_min=1e-5,
+            warmup_lr_init=1e-6,
+            t_in_epochs=True,
+        )
+        return [optimizer], [{"scheduler": scheduler, "interval": "epoch"}]
+    
+    def configure_optimizers(self):
         opt = hydra.utils.instantiate(self.hparams.optimizer, params=self.parameters(), _convert_="partial")
         if "lr_scheduler" not in self.hparams:
             return [opt]
         scheduler = hydra.utils.instantiate(self.hparams.lr_scheduler, optimizer=opt)
-        return [opt], [scheduler]
+        return [opt], [{"scheduler": scheduler, "interval": "epoch"}]
+
+    def lr_scheduler_step(self, scheduler, optimizer_idx, metric):
+      
+        scheduler.step(epoch=self.current_epoch)  # timm's scheduler need the epoch value
 
 
-@hydra.main(config_path=str(PROJECT_ROOT / "conf"), config_name="default")
+
+@hydra.main(config_path=str(PROJECT_ROOT / "conf"), config_name="default", version_base=None)
 def main(cfg: omegaconf.DictConfig) -> None:
     """Debug main to quickly develop the Lightning Module.
 
@@ -160,3 +156,32 @@ def main(cfg: omegaconf.DictConfig) -> None:
 
 if __name__ == "__main__":
     main()
+'''
+
+def main():
+    wandb.init()
+    torch.manual_seed(1)
+    dm = CIFAR10DataModule(batch_size=768)
+    model = VisionTransformer(n_classes=10)
+    lightning_model = LightningModel(model=model, learning_rate=5e-4 * 1.5)
+    trainer = L.Trainer(
+        max_epochs=1000,
+        accelerator="auto",
+        devices="auto",
+        deterministic=True,
+        logger=WandbLogger(),
+    )
+    trainer.fit(model=lightning_model, datamodule=dm)
+    train_acc = trainer.validate(dataloaders=dm.train_dataloader())
+    val_acc = trainer.validate(datamodule=dm)
+    test_acc = trainer.test(datamodule=dm)
+    print(
+        f"Train accuracy: {train_acc[0]['val acc'] * 100:.2f} | Validation accuracy: {val_acc[0]['val acc'] * 100:.2f} | Test accuracy: {test_acc[0]['test acc'] * 100:.2f}"
+    )
+
+
+if __name__ == "__main__":
+    main()
+
+
+'''
